requirements.txt
PyGithub==1.53
mypy==0.790
argparse==1.4.0
pyyaml
ruamel.yaml
